requirements.txt
alabaster==0.7.13
asttokens==2.4.0
Babel==2.12.1
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
coverage==7.3.1
debugpy==1.8.0
decorator==5.1.1
dependency-injector==4.41.0
distlib==0.3.7
docutils==0.20.1
executing==1.2.0
filelock==3.12.4
flake8==6.1.0
Flask==2.3.3
greenlet==2.0.2
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.8
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
punch.py==2.0.0
pure-eval==0.2.2
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
PyMySQL==1.1.0
pyproject-api==1.6.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.21
stack-data==0.6.2
tornado==6.3.3
tox==4.11.3
traitlets==5.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.5
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
